Revert "Add more fields for registration page"
This reverts commit 48657015a96ee3abca52aad09423cb0e6ce91ffa.
