+++ lib/constants.dart
@@ -16,14 +16,3 @@ enum Gender {
 
   const Gender(this.value);
 }
-
-enum NumberType {
-  mobile("Mobile"),
-  home("Home"),
-  work("Work"),
-  other("Other");
-
-  final String value;
-
-  const NumberType(this.value);
-}
