+++ lib/views/new_account.dart
@@ -1,0 +1,203 @@
+import 'package:flutter/material.dart';
+
+class NewAccount extends StatelessWidget {
+  const NewAccount({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final bool isSmallScreen = MediaQuery.of(context).size.width < 600;
+
+    return Scaffold(
+        body: Center(
+            child: isSmallScreen
+                ? Column(
+                    mainAxisSize: MainAxisSize.min,
+                    children: const [
+                      Logo(),
+                      FormContent(),
+                    ],
+                  )
+                : Container(
+                    padding: const EdgeInsets.all(32.0),
+                    constraints: const BoxConstraints(maxWidth: 800),
+                    child: Row(
+                      children: const [
+                        Expanded(child: Logo()),
+                        Expanded(
+                          child: Center(child: FormContent()),
+                        ),
+                      ],
+                    ),
+                  )));
+  }
+}
+
+class Logo extends StatelessWidget {
+  const Logo({Key? key}) : super(key: key);
+
+  Widget gap() => const SizedBox(height: 30);
+
+  @override
+  Widget build(BuildContext context) {
+    final bool isSmallScreen = MediaQuery.of(context).size.width < 600;
+
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Text(
+            "Create New Account",
+            textAlign: TextAlign.center,
+            style: isSmallScreen
+                ? Theme.of(context).textTheme.headline5
+                : Theme.of(context)
+                    .textTheme
+                    .headline4
+                    ?.copyWith(color: Colors.black),
+          ),
+        ),
+        gap(),
+      ],
+    );
+  }
+}
+
+class FormContent extends StatefulWidget {
+  const FormContent({Key? key}) : super(key: key);
+
+  @override
+  State<FormContent> createState() => FormContentState();
+}
+
+class FormContentState extends State<FormContent> {
+  bool isPosswordVisible = false;
+  String first = "";
+
+  final GlobalKey<FormState> formKey = GlobalKey<FormState>();
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      constraints: const BoxConstraints(maxWidth: 300),
+      child: Form(
+        key: formKey,
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            TextFormField(
+              validator: (value) {
+                // add email validation
+                if (value == null || value.isEmpty) {
+                  return 'Please enter some text';
+                }
+
+                bool emailValid = RegExp(
+                        r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
+                    .hasMatch(value);
+                if (!emailValid) {
+                  return 'Please enter a valid email';
+                }
+
+                return null;
+              },
+              decoration: const InputDecoration(
+                labelText: 'Email',
+                hintText: 'Enter your email',
+                prefixIcon: Icon(Icons.email_outlined),
+                border: OutlineInputBorder(),
+              ),
+            ),
+            gap(),
+            TextFormField(
+              validator: (value) {
+                if (value == null || value.isEmpty) {
+                  return 'Please enter some text';
+                }
+
+                if (value.length < 6) {
+                  return 'Password must be at least 6 characters';
+                }
+
+                return null;
+              },
+              obscureText: !isPosswordVisible,
+              decoration: InputDecoration(
+                  labelText: 'Password',
+                  hintText: 'Enter your password',
+                  prefixIcon: const Icon(Icons.lock_outline_rounded),
+                  border: const OutlineInputBorder(),
+                  suffixIcon: IconButton(
+                    icon: Icon(isPosswordVisible
+                        ? Icons.visibility_off
+                        : Icons.visibility),
+                    onPressed: () {
+                      setState(() {
+                        isPosswordVisible = !isPosswordVisible;
+                      });
+                    },
+                  )),
+              onChanged: (value) => {first = value},
+            ),
+            gap(),
+            TextFormField(
+              validator: (value) {
+                if (value == null || value.isEmpty) {
+                  return 'Please enter some text';
+                }
+
+                if (value != first) {
+                  return 'Password does not match';
+                }
+                return null;
+              },
+              obscureText: !isPosswordVisible,
+              decoration: InputDecoration(
+                  labelText: 'Verify Password',
+                  hintText: 'Verify Password',
+                  prefixIcon: const Icon(Icons.lock_outline_rounded),
+                  border: const OutlineInputBorder(),
+                  suffixIcon: IconButton(
+                    icon: Icon(isPosswordVisible
+                        ? Icons.visibility_off
+                        : Icons.visibility),
+                    onPressed: () {
+                      setState(() {
+                        isPosswordVisible = !isPosswordVisible;
+                      });
+                    },
+                  )),
+            ),
+            gap(),
+            SizedBox(
+              width: double.infinity,
+              child: ElevatedButton(
+                style: ElevatedButton.styleFrom(
+                  shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(4)),
+                ),
+                child: const Padding(
+                  padding: EdgeInsets.all(10.0),
+                  child: Text(
+                    'Create New Account',
+                    style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+                  ),
+                ),
+                onPressed: () {
+                  // if (formKey.currentState?.validate() ?? false) {
+                  //   debugPrint("Validateed");
+                  //   // Create new account and insert data into database
+                  // }
+                  Navigator.pushNamed(context, "/provider");
+                },
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  Widget gap() => const SizedBox(height: 16);
+}
